feat: Implement unary signs optimizers
Implement unary signs optimizers to simplify expressions by removing unary operations like -x or +x. This optimizer can produce complex token sequences and should be used with other optimizers like brackets and constant folding optimizers for maximum effectiveness.

Example: A subexpression ‘-(-(-a))’ will be converted to '(0 - ((0 - (0 - a))))' using this optimizer, resulting in the same expression but in a different representation.